mitigation sim
